approvals: extract terminate_approval clipping into utils
