+++ itou/approvals/utils.py
@@ -1,11 +1,15 @@
 import datetime
+import logging
 
 from django.db.models import Max, Q
 from django.utils import timezone
 
 from itou.companies.enums import CompanyKind
 from itou.companies.models import Contract
+from itou.utils.admin import add_support_remark_to_obj
 
+
+logger = logging.getLogger(__name__)
 
 # A PASS IAE is eligible for closure when its suspension has been going on for
 # at least this duration without interruption.
@@ -13,7 +17,7 @@ SUSPENSION_DURATION_BEFORE_APPROVAL_CLOSABLE = datetime.timedelta(days=365)
 
 
 def can_close_approval(approval):
-    """Return True when *approval* meets all three conditions for a user-initiated closure:
+    """Return True when approval meets all three conditions for a user-initiated closure:
 
     1. At least one suspension has been running (or ran) for more than 12
        consecutive months, and no accepted hiring occurred after it ended.
@@ -55,6 +59,59 @@ def can_close_approval(approval):
         Q(end_date__isnull=True) | Q(end_date__gte=today),
         job_seeker=approval.user,
     ).exists()
+
+
+def _clip_approval_dependency(approval, model, end_date, acting_user):
+    _, deletions = model.objects.filter(approval=approval, start_at__gte=end_date).delete()
+    if deletions:
+        logger.info(
+            "Terminating approval pk=%(approval_id)d, deleting %(deletions)d future %(model_name)s.",
+            {
+                "approval_id": approval.pk,
+                "deletions": deletions[model._meta.label],
+                "model_name": model._meta.label,
+            },
+        )
+    try:
+        obj = model.objects.in_progress().filter(approval=approval).get()
+    except model.DoesNotExist:
+        pass
+    else:
+        logger.info(
+            "Terminating approval pk=%(approval_id)d, "
+            "setting %(model_name)s pk=%(model_id)d end_at=%(end_at)s "
+            "(was %(initial_end_at)s).",
+            {
+                "approval_id": approval.pk,
+                "model_name": obj._meta.label,
+                "model_id": obj.pk,
+                "end_at": end_date,
+                "initial_end_at": obj.end_at,
+            },
+        )
+        obj.end_at = end_date
+        obj.updated_by = acting_user
+        obj.save(update_fields=["end_at", "updated_at", "updated_by"])
+
+
+def close_approval(approval, *, closed_by):
+    """Terminate approval as of today, clipping its ongoing suspensions and prolongations."""
+    from itou.approvals.models import Prolongation, Suspension
+
+    new_end = timezone.localdate()
+    _clip_approval_dependency(approval, Prolongation, new_end, closed_by)
+    _clip_approval_dependency(approval, Suspension, new_end, closed_by)
+    logger.info(
+        "Terminating approval pk=%(approval_id)d, end_at=%(end_at)s (was %(initial_end_at)s).",
+        {
+            "approval_id": approval.pk,
+            "initial_end_at": approval.end_at,
+            "end_at": new_end,
+        },
+    )
+    approval.end_at = new_end
+    approval.save(update_fields=["end_at", "updated_at"])
+    add_support_remark_to_obj(approval, f"{new_end} : PASS IAE clôturé par {closed_by.get_full_name()}.")
 
 
 def get_user_last_accepted_siae_job_application(user):
